+++ src/main/java/com/cesi/product/ProductService.java
@@ -1,5 +1,6 @@
 package com.cesi.product;
 
+import com.cesi.process.Process;
 import com.cesi.ProductIngredient;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.transaction.Transactional;
@@ -34,17 +35,35 @@ public class ProductService {
 
     public Product updateProduct(@Valid Product product) {
         Product entity = Product.findById(product.id);
-        entity.nom = product.nom;
-        entity.description = product.description;
-        entity.pUHT = product.pUHT;
-        entity.gamme = product.gamme;
+        if (entity != null) {
+            // Supprimer toutes les occurrences de ProductIngredient associées à ce Product
+            ProductIngredient.delete("product = ?1", entity);
 
-        entity.productIngredients = product.productIngredients;
+            // Mettre à jour les propriétés du Product
+            entity.nom = product.nom;
+            entity.description = product.description;
+            entity.pUHT = product.pUHT;
+            entity.gamme = product.gamme;
+
+            // Assurez-vous que les références sont correctement établies
+            for (ProductIngredient productIngredient : product.productIngredients) {
+                productIngredient.product = entity;
+            }
+
+            // Enregistrez les modifications
+            entity.persist();
+        }
         return entity;
     }
 
     public void deleteProduct(Long id) {
         Product product = Product.findById(id);
-        product.delete();
+        if (product != null) {
+            Process.delete("product = ?1", product);
+            for (ProductIngredient productIngredient : product.productIngredients) {
+                productIngredient.delete();
+            }
+            product.delete();
+        }
     }
 }
